requirements.txt
discord.py
twisted-fate.py